requirements.txt
discord.py-self[voice]
yt-dlp